Simplify and update lint
