+++ .github/workflows/lint.yml
@@ -11,7 +11,7 @@ jobs:
     - uses: actions/checkout@v3
     - uses: actions/setup-go@v4
       with:
-        go-version: 1.20.x
+        go-version: 1.21.x
     - run: go version
     - name: Run golangci-lint
       uses: golangci/golangci-lint-action@v3
